resolved layout issues, slimmed down code
