+++ packages/components/app/styles/components/application-state.scss
@@ -6,47 +6,34 @@
 //
 // APPLICATION-STATE COMPONENT
 //
+
 $hds-application-state-padding: 12px 0;
 
 .hds-application-state {
-  align-self: center;
-  width: 20rem;
+  width: 19.5rem;
   max-width: 100%;
-  margin: 0 auto;
-}
-
-.hds-application-state__body,
-.hds-application-state__footer {
-  display: flex;
+  margin: 0 auto; // this will center the component in the parent container
 }
 
 .hds-application-state__header {
   display: grid;
-  grid-row-gap: 0;
-  grid-column-gap: 0;
+  grid-template-columns: min-content 1fr;
+  align-items: start;
   color: var(--token-color-foreground-primary);
+}
 
-  .hds-application-state--error & {
-    grid-template-rows: repeat(2, auto);
-    grid-template-columns: 28px auto;
-    align-items: stretch;
+.hds-application-state__icon {
+  margin-right: 8px; // instead of column gap
+  padding-top: 4px; // this seems to align the icon along with line 21
+}
 
-    .hds-application-state__header-icon {
-      grid-area: 1 / 1 / 3 / 2;
-      padding-top: 4px;
-    }
-
-    .hds-application-state__header-title {
-      grid-area: 1 / 2 / 2 / 3;
-    }
-
-    .hds-application-state__header-errorCode {
-      grid-area: 2 / 2 / 3 / 3;
-    }
-  }
+.hds-application-state__title,
+.hds-application-state__error-code {
+  grid-column-start: 2;
 }
 
 .hds-application-state__body {
+  display: flex;
   padding: $hds-application-state-padding;
   color: var(--token-color-foreground-faint);
 }
@@ -55,8 +42,14 @@ $hds-application-state-padding: 12px 0;
   display: flex;
   justify-content: space-between;
 
-  .hds-application-state--has-divider & {
+  &.hds-application-state__footer--has-divider {
     padding: $hds-application-state-padding;
     border-top: 1px solid var(--token-color-border-strong);
   }
+
+  // in case users do something outside design guidelines
+  .hds-link-standalone + .hds-link-standalone {
+    margin-left: 8px;
+  }
 }
+
