+++ website/docs/components/application-state/index.md
@@ -1,18 +1,17 @@
 ---
 title: Application State
-description: Use the ApplicationState to provide information about the current state of the application.
-caption: Use the ApplicationState to provide information about the current state of the application.
+description: An informational element that displays the current state of the application.
+caption: An informational element that displays the current state of the application.
 status: released
 links:
-  figma: https://www.figma.com/file/noyY6dUMDYjmySpHcMjhkN/branch/2lNa9Qe4UuF4yN8gcR7Fkp/HDS-Product---Components?node-id=31113%3A50004&t=AgcOxZqO6FWFkdcD-1
+  figma: https://www.figma.com/file/noyY6dUMDYjmySpHcMjhkN/HDS-Product---Components?node-id=31113-50004
   github: https://github.com/hashicorp/design-system/tree/main/packages/components/addon/components/hds/application-state
 previewImage: assets/illustrations/components/application-state.jpg
 navigation:
-  keywords: ['empty state', 'error state', 'application state', 'message']
+  keywords: ['empty state', 'error state', 'message']
 ---
 
 <section data-tab="Guidelines">
-  @include "partials/guidelines/overview.md"
   @include "partials/guidelines/guidelines.md"
 </section>
