+++ website/docs/components/application-state/partials/code/component-api.md
@@ -1,12 +1,6 @@
 ## Component API
 
 <Doc::ComponentApi as |C|>
-  <C.Property @name="errorCode" @type="string">
-    The error code to be displayed.
-  </C.Property>
-  <C.Property @name="hasDivider" @type="boolean" @default="false" @values={{array "true" "false"}}>
-    Indicates if there should be a visible divider above the footer.
-  </C.Property>
   <C.Property @name="...attributes">
     This component supports use of [`...attributes`](https://guides.emberjs.com/release/in-depth-topics/patterns-for-components/#toc_attribute-ordering).
   </C.Property>
@@ -17,15 +11,18 @@
 #### [A].Header
 
 <Doc::ComponentApi as |C|>
+  <C.Property @name="errorCode" @type="string">
+    The error code to be displayed.
+  </C.Property>
   <C.Property @name="icon" @type="string">
-    Shows an icon if `@errorCode` is defined when the component is invoked. Default icon is `alert-circle`, but any valid [icon](/icons/library) name is supported.
+    Adds a leading icon to the title. Accepts any [icon](/icons/library) name.
   </C.Property>
   <C.Property @name="title" @type="string"  />
 </Doc::ComponentApi>
 
 #### [A].Body
 
-Supports block use.
+Supports block invocation for custom use.
 
 <Doc::ComponentApi as |C|>
   <C.Property @name="text" @type="string" />
@@ -33,4 +30,10 @@ Supports block use.
   
 #### [A].Footer
 
-Supports block use; yields the [Link::Standalone](../components/link/standalone) component as `Link`.
+Yields the [Link::Standalone](../components/link/standalone) component as `Link::Standalone`.
+
+<Doc::ComponentApi as |C|>
+  <C.Property @name="hasDivider" @type="boolean" @default="false" @values={{array "true" "false"}}>
+    Indicates if there should be a visible divider above the footer.
+  </C.Property>
+</Doc::ComponentApi>
