+++ website/docs/components/application-state/partials/code/how-to-use.md
@@ -1,12 +1,21 @@
 ## How to use this component
 
-This component currently supports two states: empty state and error state.
+This component intends to replace a few different simple error and empty/zero state components that exist across product UIs, while providing flexibility to adapt to additional application states. While we’re referring to these as "error state" and "empty state" this is just to help consumers relate to the components that have been replaced. Really, this component can be used to reflect any kind of application state message that is needed.
 
-### Empty state
+### As an empty state
 
 ```handlebars
 <Hds::ApplicationState as |A|>
   <A.Header @title="Empty state title text" />
+  <A.Body @text="The item you were looking for was not found." />
+</Hds::ApplicationState>
+```
+
+#### Empty state with header icon
+
+```handlebars
+<Hds::ApplicationState as |A|>
+  <A.Header @title="Empty state title text" @icon="alert-circle" />
   <A.Body @text="The item you were looking for was not found." />
 </Hds::ApplicationState>
 ```
@@ -18,7 +27,7 @@ This component currently supports two states: empty state and error state.
   <A.Header @title="Empty state title text" />
   <A.Body @text="The item you were looking for was not found." />
   <A.Footer as |F|>
-    <F.Link @icon="help" @text="Need Help" @href="/components/alert"
+    <F.Link::Standalone @icon="help" @text="Need Help" @href="/components/alert"
     @iconPosition="trailing" />
   </A.Footer>
 </Hds::ApplicationState>
@@ -27,11 +36,11 @@ This component currently supports two states: empty state and error state.
 #### Empty state with a footer link and divider
 
 ```handlebars
-<Hds::ApplicationState @hasDivider={{true}} as |A|>
+<Hds::ApplicationState as |A|>
   <A.Header @title="Empty state title text" />
   <A.Body @text="The item you were looking for was not found." />
-  <A.Footer as |F|>
-    <F.Link @icon="help" @text="Need Help" @href="/components/alert"
+  <A.Footer @hasDivider={{true}} as |F|>
+    <F.Link::Standalone @icon="help" @text="Need Help" @href="/components/alert"
     @iconPosition="trailing" />
   </A.Footer>
 </Hds::ApplicationState>
@@ -46,42 +55,42 @@ This component currently supports two states: empty state and error state.
     <Doc::Placeholder @text="block yield" @height="100" @background="#eee" />
   </A.Body>
   <A.Footer as |F|>
-    <F.Link @icon="arrow-left" @text="Go back" @href="/" />
+    <F.Link::Standalone @icon="arrow-left" @text="Go back" @href="/" />
   </A.Footer>
 </Hds::ApplicationState>
 ```
 
-### Error state
+### As an error state
 
-To indicate that the message is an error state, add `@errorCode` to the invocation.
+To indicate that the message is an error state, add `@errorCode` to the `[A].Header` component invocation.
 
 ```handlebars
-<Hds::ApplicationState @errorCode="404" as |A|>
-  <A.Header @title="An error has occurred" />
+<Hds::ApplicationState as |A|>
+  <A.Header @title="An error has occurred" @errorCode="404" />
   <A.Body
     @text="Sorry, an unexpected error has occurred.
     Please try again later or contact support for assistance."
   />
   <A.Footer as |F|>
-    <F.Link @icon="arrow-left" @text="Go back" @href="/" />
-    <F.Link @icon="help" @text="Need Help" @href="/components/alert" 
+    <F.Link::Standalone @icon="arrow-left" @text="Go back" @href="/" />
+    <F.Link::Standalone @icon="help" @text="Need Help" @href="/components/alert" 
       @iconPosition="trailing" />
   </A.Footer>
 </Hds::ApplicationState>
 ```
 
-#### Error state with customized header icon
+#### Error state with header icon
 
 ```handlebars
-<Hds::ApplicationState @errorCode="404" as |A|>
-  <A.Header @title="An error has occurred" @icon="help" />
+<Hds::ApplicationState as |A|>
+  <A.Header @title="An error has occurred" @icon="help" @errorCode="404" />
   <A.Body
     @text="Sorry, an unexpected error has occurred.
     Please try again later or contact support for assistance."
   />
   <A.Footer as |F|>
-    <F.Link @icon="arrow-left" @text="Go back" @href="/" />
-    <F.Link @icon="help" @text="Need Help" @href="/components/alert" 
+    <F.Link::Standalone @icon="arrow-left" @text="Go back" @href="/" />
+    <F.Link::Standalone @icon="help" @text="Need Help" @href="/components/alert" 
       @iconPosition="trailing" />
   </A.Footer>
 </Hds::ApplicationState>
@@ -90,14 +99,14 @@ To indicate that the message is an error state, add `@errorCode` to the invocati
 #### Error state with yielded body block
 
 ```handlebars
-<Hds::ApplicationState @errorCode="404" as |A|>
-  <A.Header @title="An error has occurred" />
+<Hds::ApplicationState as |A|>
+  <A.Header @title="An error has occurred" @errorCode="404" />
   <A.Body>
     <Doc::Placeholder @text="block yield" @height="100" @background="#eee" />
   </A.Body>
   <A.Footer as |F|>
-    <F.Link @icon="arrow-left" @text="Go back" @href="/" />
-    <F.Link @icon="help" @text="Need Help" 
+    <F.Link::Standalone @icon="arrow-left" @text="Go back" @href="/" />
+    <F.Link::Standalone @icon="help" @text="Need Help" 
       @href="/components/alert" @iconPosition="trailing" />
   </A.Footer>
 </Hds::ApplicationState>
@@ -106,15 +115,15 @@ To indicate that the message is an error state, add `@errorCode` to the invocati
 #### Error state with a footer divider
 
 ```handlebars
-<Hds::ApplicationState @errorCode="404" @hasDivider={{true}} as |A|>
-  <A.Header @title="An error has occurred" />
+<Hds::ApplicationState as |A|>
+  <A.Header @title="An error has occurred" @errorCode="404" />
   <A.Body
     @text="Sorry, an unexpected error has occurred.
     Please try again later or contact support for assistance."
   />
-  <A.Footer as |F|>
-    <F.Link @icon="arrow-left" @text="Go back" @href="/" />
-    <F.Link @icon="help" @text="Need Help" @href="/components/alert" 
+  <A.Footer @hasDivider={{true}} as |F|>
+    <F.Link::Standalone @icon="arrow-left" @text="Go back" @href="/" />
+    <F.Link::Standalone @icon="help" @text="Need Help" @href="/components/alert" 
       @iconPosition="trailing" />
   </A.Footer>
 </Hds::ApplicationState>
