+++ website/docs/components/application-state/partials/guidelines/guidelines.md
@@ -1,25 +1,29 @@
 ## Usage
 
 ### When to use
+
 - When a user's action does not yield any results due to incorrect or incomplete input.
 - When an application encounters an issue or error during its operation.
 
 ### When not to use
-- When the absence of content is expected and does not require an explanation to the user. 
-- When there is a clear and intuitive way to add or populate content. 
+
+- When the absence of content is expected and does not require an explanation to the user.
+- When there is a clear and intuitive way to add or populate content.
 
 ## Icon
 
-- The icon should be relevant and communicate the same purpose as the errorCode.
+- The icon should align with the purpose of the content and effectively communicate the same message.
 
 ## Actions
 
-- In the footer, you can include two stand-alone links. 
-- It’s important to keep these links separate from call-to-action buttons (CTA), as doing so may cause misalignment with the code.
+- In the footer, you can include up to two stand-alone links. 
+- We don't recommend using buttons, as most actions will navigate the user away from this page. Learn more about [when to use a link vs. a button](https://helios.hashicorp.design/components/link/standalone#usage).
+- For the standalone link, we recommend using the medium size.
+- Use footer actions to redirect or guide users in solving errors/access issues with actionable steps.
 
 ## Content
 
-- The title should be short of providing a clear and concise message. 
+- The title should be short and provide a clear and concise message.
 - Focus on relevant information and avoid unnecessary details.
 - Provide a straightforward explanation of the problem or error.
-- Suggestions or guidance for how the user can resolve the issue, if possible. 
+- Include suggestions or guidance for how the user can resolve the issue, if possible.
