--- website/docs/components/application-state/partials/guidelines/overview.md
@@ -1,1 +1,0 @@
-Use footer actions to redirect or guide users in solving errors/access issues with actionable steps.
